requirements.txt
anyio==4.4.0
ase==3.23.0
astor==0.8.1
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.6.2
cfgv==3.4.0
charset-normalizer==3.3.2
chemparse==0.1.0
colorlog==6.8.2
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
dnspython==2.6.1
docopt==0.6.2
einops==0.8.0
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
fonttools==4.53.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.5.36
idna==3.7
importlib_metadata==8.0.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
latexcodec==3.0.0
MarkupSafe==2.1.5
matminer==0.9.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mistune==3.0.2
monty==2024.5.24
mpmath==1.3.0
multiprocess==0.70.16
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
nodeenv==1.9.1
numpy==1.26.4
opt-einsum==3.3.0
p_tqdm==1.4.0
packaging==24.1
paconvert==1.0.0
paddlepaddle-gpu==2.6.1
palettable==3.3.3
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathos==0.3.2
pexpect==4.9.0
pgl==2.2.3
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.2.2
plotly==5.22.0
pox==0.3.4
ppft==1.7.6.8
pre-commit==3.7.1
prompt_toolkit==3.0.47
protobuf==5.27.1
ptyprocess==0.7.0
pure-eval==0.2.2
pybtex==0.24.0
Pygments==2.18.0
pymatgen==2024.6.10
pymongo==4.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyxtal==0.6.7
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
remote-pdb==2.1.0
requests==2.32.3
rpds-py==0.18.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
SMACT==2.6
sniffio==1.3.1
soupsieve==2.5
spglib==2.4.0
stack-data==0.6.3
sympy==1.12.1
tabulate==0.9.0
tenacity==8.4.0
threadpoolctl==3.5.0
tinycss2==1.3.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uncertainties==3.2.1
urllib3==2.2.1
virtualenv==20.26.3
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
zipp==3.19.2
